Add support for YAML options in charts, and (by reusing the same config handler as %%bigquery) include support for variables, which is useful as we now parse the JSON rather than just passing it on to Charts and so cannot include expressions any more.
